+++ src/app/repository/PeopleRepository.js
@@ -1,5 +1,5 @@
 const PeopleSchema = require('../schema/PeopleSchema')
-
+const PeopleIdNotFound = require('../errors/people/PeopleIdNotFound')
 
 class PeopleRepository {
     async create(payload) {
@@ -13,14 +13,25 @@ class PeopleRepository {
     }
 
     async findOneById(payload) {
-        return PeopleSchema.findById({ _id: payload })
+        const finded = await PeopleSchema.findById({ _id: payload })
+        if (finded == null) {
+            throw new PeopleIdNotFound(payload)
+        }
+        return finded
+    }
+
+    async UpdateOneById(id, payload) {
+        const finded = PeopleSchema.findOneAndUpdate({ _id: id }, payload, { returnOriginal: false })
+        return finded
     }
 
     async deleteOne(payload) {
+        const finded = await PeopleSchema.findById({ _id: payload })
+        if (finded == null) {
+            throw new PeopleIdNotFound(payload)
+        }
         return PeopleSchema.deleteOne({ _id: payload })
     }
 }
-
-
 
 module.exports = new PeopleRepository()
